+++ lib/src/xml/navigation/ancestors.dart
@@ -1,11 +1,17 @@
 import 'dart:collection';
 
+import '../nodes/element.dart';
 import '../nodes/node.dart';
 
 extension XmlAncestorsExtension on XmlNode {
-  /// Return a lazy [Iterable] of the ancestors of this [XmlNode] (parent,
+  /// Return a lazy [Iterable] of the ancestors of this node (parent,
   /// grandparent, ...) in reverse document order.
   Iterable<XmlNode> get ancestors => XmlAncestorsIterable(this);
+
+  /// Return a lazy [Iterable] of the ancestor [XmlElement] nodes of this
+  /// node (parent, grandparent, ...) in reverse document order.
+  Iterable<XmlElement> get ancestorElements =>
+      ancestors.whereType<XmlElement>();
 }
 
 /// Iterable to walk over the ancestors of a node.