+++ lib/src/xml/navigation/following.dart
@@ -1,12 +1,18 @@
 import 'dart:collection';
 
 import '../nodes/attribute.dart';
+import '../nodes/element.dart';
 import '../nodes/node.dart';
 
 extension XmlFollowingExtension on XmlNode {
-  /// Return a lazy [Iterable] of the nodes following this [XmlNode] in
-  /// document order.
+  /// Return a lazy [Iterable] of the nodes following this node in document
+  /// order.
   Iterable<XmlNode> get following => XmlFollowingIterable(this);
+
+  /// Return a lazy [Iterable] of the [XmlElement] nodes following this node
+  /// in document order.
+  Iterable<XmlElement> get followingElements =>
+      following.whereType<XmlElement>();
 }
 
 /// Iterable to walk over the followers of a node.