+++ test/navigation_test.dart
@@ -28,6 +28,19 @@ void main() {
     expect(book.children[0].children[1].ancestors, [book.children[0], book]);
     verifyIterator(book.children[0].children[1].ancestors);
   });
+  test('ancestorElements', () {
+    expect(book.ancestorElements, []);
+    expect(book.children[0].ancestorElements, []);
+    expect(book.children[0].children[0].ancestorElements, [book.children[0]]);
+    expect(book.children[0].children[0].attributes[0].ancestorElements,
+        [book.children[0].children[0], book.children[0]]);
+    expect(book.children[0].children[0].attributes[1].ancestorElements,
+        [book.children[0].children[0], book.children[0]]);
+    expect(book.children[0].children[0].children[0].ancestorElements,
+        [book.children[0].children[0], book.children[0]]);
+    expect(book.children[0].children[1].ancestorElements, [book.children[0]]);
+    verifyIterator(book.children[0].children[1].ancestorElements);
+  });
   test('preceding', () {
     expect(book.preceding, []);
     expect(book.children[0].preceding, [book]);
@@ -57,6 +70,20 @@ void main() {
     ]);
     verifyIterator(book.children[0].children[1].preceding);
   });
+  test('precedingElements', () {
+    expect(book.precedingElements, []);
+    expect(book.children[0].precedingElements, []);
+    expect(book.children[0].children[0].precedingElements, [book.children[0]]);
+    expect(book.children[0].children[0].attributes[0].precedingElements,
+        [book.children[0], book.children[0].children[0]]);
+    expect(book.children[0].children[0].attributes[1].precedingElements,
+        [book.children[0], book.children[0].children[0]]);
+    expect(book.children[0].children[0].children[0].precedingElements,
+        [book.children[0], book.children[0].children[0]]);
+    expect(book.children[0].children[1].precedingElements,
+        [book.children[0], book.children[0].children[0]]);
+    verifyIterator(book.children[0].children[1].precedingElements);
+  });
   test('descendants', () {
     expect(book.descendants, [
       book.children[0],
@@ -84,6 +111,21 @@ void main() {
     expect(book.children[0].children[1].descendants, []);
     verifyIterator(book.descendants);
   });
+  test('descendantElements', () {
+    expect(book.descendantElements, [
+      book.children[0],
+      book.children[0].children[0],
+      book.children[0].children[1]
+    ]);
+    expect(book.children[0].descendantElements,
+        [book.children[0].children[0], book.children[0].children[1]]);
+    expect(book.children[0].children[0].descendantElements, []);
+    expect(book.children[0].children[0].attributes[0].descendantElements, []);
+    expect(book.children[0].children[0].attributes[1].descendantElements, []);
+    expect(book.children[0].children[0].children[0].descendantElements, []);
+    expect(book.children[0].children[1].descendantElements, []);
+    verifyIterator(book.descendantElements);
+  });
   test('following', () {
     expect(book.following, []);
     expect(book.children[0].following, []);
@@ -102,6 +144,20 @@ void main() {
         [book.children[0].children[1]]);
     expect(book.children[0].children[1].following, []);
     verifyIterator(book.following);
+  });
+  test('followingElements', () {
+    expect(book.followingElements, []);
+    expect(book.children[0].followingElements, []);
+    expect(book.children[0].children[0].followingElements,
+        [book.children[0].children[1]]);
+    expect(book.children[0].children[0].attributes[0].followingElements,
+        [book.children[0].children[1]]);
+    expect(book.children[0].children[0].attributes[1].followingElements,
+        [book.children[0].children[1]]);
+    expect(book.children[0].children[0].children[0].followingElements,
+        [book.children[0].children[1]]);
+    expect(book.children[0].children[1].followingElements, []);
+    verifyIterator(book.followingElements);
   });
   test('nodes', () {
     expect(book.nodes, [book.children[0]]);