+++ lib/src/xml/utils/attribute_type.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: constant_identifier_names
+
 /// Enum of the attribute quote types.
 enum XmlAttributeType {
   SINGLE_QUOTE,