+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## 5.2.0
 
-* Add `XmlNode.nextElementSibling` and `XmlNode.previousElementSibling`. 
-* Add `XmlNode.siblings` and `XmlNode.siblingElements`.
-* Add `XmlNode.childElements`.
+* A series of accessors that simplify navigating the XML DOM with `XmlElements`:
+  * Add `XmlNode.childElements`.
+  * Add `XmlNode.siblings` and `XmlNode.siblingElements`.
+  * Add `XmlNode.nextElementSibling` and `XmlNode.previousElementSibling`. 
+  * Add `XmlNode.ancestorElements`, `XmlNode.precedingElements`, `XmlNode.descendantElements`, and `XmlNode.followingElements`.
 
 ## 5.1.1